% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap an renv Installation}
\usage{
bootstrap(project = NULL, version = NULL)
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active project will
be used. If no project is currently active, then the current working
directory is used instead.}

\item{version}{The version of \code{renv} to install. If \code{NULL}, the version
of \code{renv} currently installed will be used. The requested version of
\code{renv} will be retrieved from the \code{renv} public GitHub repository,
at \url{https://github.com/rstudio/renv}.}
}
\value{
The project directory, invisibly. Note that this function is normally
called for its side effects.
}
\description{
Bootstrap an \code{renv} installation, making the requested version of
\code{renv} available for projects on the system.
}
\details{
Normally, this function does not need to be called directly by the user; it
will be invoked as required by \code{\link[=init]{init()}} and \code{\link[=activate]{activate()}}.
}
