% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite.R
\name{cite}
\alias{cite}
\title{Generate Citations for Packages used in a Project}
\usage{
cite(type = c("plain", "bibtex"), ..., project = NULL)
}
\arguments{
\item{type}{The type of citation(s) to generate.}
}
\description{
Generate Citations for Packages used in a Project
}
