% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean a Project}
\usage{
clean(project = NULL, confirm = interactive())
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active
project will be used. If no project has been specifically activated,
the current working directory is used.}

\item{confirm}{Boolean; prompt the user before taking any action?}
}
\description{
Clean up a project and its associated \R libraries.
}
\details{
The following actions will be performed:
\itemize{
\item Leftover temporary directories in the project library will be removed.
\item Non-system packages installed in the system library will be removed.
}
}
