% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{deactivate}
\alias{deactivate}
\title{Deactivate a Project}
\usage{
deactivate(project = NULL)
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active
project will be used. If no project has been specifically activated,
the current working directory is used.}
}
\description{
Use \code{deactivate()} to remove the infrastructure used by \code{renv} to activate
projects for newly-launched \R sessions.
}
\seealso{
Other renv: \code{\link{activate}}
}
\concept{renv}
