% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\title{Print a Diagnostics Report}
\usage{
diagnostics(project = NULL)
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active project will
be used. If no project is currently active, then the current working
directory is used instead.}
}
\value{
This function is normally called for its side effects.
}
\description{
Print a diagnostics report, summarizing the state of a project using \code{renv}.
This report can occasionally be useful when diagnosing issues with \code{renv}.
}
