% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{discover_dependencies}
\alias{discover_dependencies}
\title{Discover R Package Dependencies}
\usage{
discover_dependencies(path = getwd())
}
\arguments{
\item{path}{The path to a (possibly multi-mode) \R file, or a directory
containing such files.}
}
\description{
Discover \R packages used within files and directories.
}
