% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equip.R
\name{equip}
\alias{equip}
\title{Install Required System Libraries}
\usage{
equip()
}
\description{
Equip your system with libraries commonly-used during compilation of
\R packages. Currently only supported on Windows.
}
