% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fork.R
\name{fork}
\alias{fork}
\title{Fork a Virtual Environment}
\usage{
fork(name)
}
\arguments{
\item{name}{The name of a virtual environment. If \code{NULL}, then the active
virtual environment (if any) will be used.}
}
\description{
Fork a (global) virtual environment into a project-local one. This is useful
for isolating a project -- e.g. you need to preserve the current state of
a global environment locally, and make your own (persistent) modifications
to that environment without affecting the original environment.
}
