% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{history}
\alias{history}
\title{View Lockfile History}
\usage{
history(project = NULL)
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active project will
be used. If no project is currently active, then the current working
directory is used instead.}
}
\description{
Use your version control system to find prior versions of the \code{renv.lock}
file that have been used in your project.
}
\details{
The \code{history()} function is currently only implemented for projects using
\code{git} for version control.
}
