% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize a Project-local Virtual Environment}
\usage{
init(project = NULL, ...)
}
\arguments{
\item{project}{The project directory.}

\item{...}{Optional arguments passed to \code{\link[=create]{create()}}.}
}
\description{
Discover packages used within the current project, and then initialize
a virtual environment with those packages.
}
