% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash.R
\name{load_vars}
\alias{load_vars}
\title{Load environment variables into the current R session}
\usage{
load_vars(bash_file)
}
\arguments{
\item{bash_file}{A connection object or character string to pass to \code{readLines}.}
}
\description{
\code{load_vars} reads the contents of a bash file, loads,
evaluates and sets environment variables in the current R session.
}
