% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{manifest}
\alias{manifest}
\title{Manifests}
\description{
A \strong{manifest} records the state of a virtual environment at some point in
time. Virtual environments can be restored from a manifest using the
\code{\link[=restore]{restore()}} function. This implies re-creating the environment, and
reinstalling any \R packages used within the libraries for that environment.
}
\details{
While manifest files are normally generated and used with \code{\link[=snapshot]{snapshot()}} /
\code{\link[=restore]{restore()}}, they can also be generated by hand and edited if so desired.
The structure is similar to that of the Windows-style \code{.ini} file, with
some provisioning for nested sections.

An example manifest file folllows:\preformatted{[Environment]
Environment=renv-994682bcc73
Version=0.1.0
Local=FALSE

[R]
Version=3.5.1
Library=dev
Overlay=FALSE
Repositories=
  CRAN=https://cran.rstudio.com

[R.Package.markdown]
Package=markdown
Version=0.9
Library=dev
Source=cran
Hash=8515151150d7372bc76e0af15ef5dee0

[R.Package.mime]
Package=mime
Version=0.6
Library=dev
Source=cran
Hash=b1e49df8aef896bc8c0b749ef1da5a48
}
}
\section{Environment}{


Base properties for the virtual environment.

\tabular{ll}{
\strong{Environment} \tab The environment name. \cr
\strong{Version}     \tab The version of the \code{renv} package used with this environment. \cr
\strong{Local}       \tab Whether this is a local or global virtual environment. \cr
}
}

\section{R}{


Properties related to the version of \R associated with this virtual environment.

\tabular{ll}{
\strong{Version}      \tab The version of \R used. \cr
\strong{Library}      \tab The names of the \R libraries used in this virtual environment. \cr
\strong{Overlay}      \tab Whether or not the user library is used (overlaid) in this virtual environment. \cr
\strong{Repositories} \tab The \R repositories used in this project.
}
}

\section{R.Package}{


Package records, related to the version of an \R package that was installed
at the time the manifest was generated.

\tabular{ll}{
\strong{Package}      \tab The package name. \cr
\strong{Version}      \tab The package version. \cr
\strong{Library}      \tab The library this package was installed in. \cr
\strong{Source}       \tab The location from which this package was retrieved. \cr
\strong{Hash}         \tab (Optional) A unique hash for this package, used for package caching. \cr
}
}

\seealso{
Other reproducibility: \code{\link{restore}},
  \code{\link{snapshot}}
}
\concept{reproducibility}
