% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{migrate}
\alias{migrate}
\title{Migrate a Project from Packrat to renv}
\usage{
migrate(project = NULL)
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active project will
be used. If no project is currently active, then the current working
directory is used instead.}
}
\description{
Migrate a project's infrastructure from Packrat to \code{renv}.
The following actions are taken:
}
\details{
\itemize{
\item The Packrat lockfile is migrated to \code{renv.lock};
\item Packages installed in the Packrat library are migrated to the \code{renv} library;
\item Relevant Packrat options (e.g. \code{ignored.packages}) are copied;
\item Packages in the Packrat cache are imported into the \code{renv} cache;
\item The project \code{.Rprofile} is updated to use \code{renv}.
}
}
