% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Retrieve the Active Project}
\usage{
project(default = NULL)
}
\arguments{
\item{default}{The value to return when no project is
currently active. Defaults to \code{NULL}.}
}
\value{
The active project directory, as a length-one character vector.
}
\description{
Retrieve the path to the active project (if any).
}
\examples{
\dontrun{

# get the currently-active renv project
renv::project()

}
}
