% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge.R
\name{purge}
\alias{purge}
\title{Purge Packages from the Cache}
\usage{
purge(package, version = NULL, hash = NULL, confirm = interactive())
}
\arguments{
\item{package}{A single package to be removed from the cache.}

\item{version}{The package version to be removed. When \code{NULL}, all versions
of the requested package will be removed.}

\item{hash}{The specific hashes to be removed. When \code{NULL}, all hashes
associated with a particular package's version will be removed.}

\item{confirm}{Boolean; prompt the user before taking any action?}
}
\description{
Purge packages from the cache. This can be useful if a package which had
previously been installed in the cache has become corrupted or unusable,
and needs to be reinstalled.
}
