% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\docType{class}
\name{registry-class}
\alias{registry-class}
\alias{registry}
\title{A registry contains a list of all currently defined environment variables,
with utilities for evaluating references to other variables.  Not meant
for external usage.}
\description{
A registry contains a list of all currently defined environment variables,
with utilities for evaluating references to other variables.  Not meant
for external usage.
}
\section{Fields}{

\describe{
\item{\code{vars}}{A named list of environment variables.}

\item{\code{export_pattern}}{Regular expression for finding export statements in environment files.}

\item{\code{var_pattern}}{Regular expression for identifying variable references.}

\item{\code{variable_pattern}}{Regular expression for finding variable declarations.}

\item{\code{comment_pattern}}{Regular expression for identifying comments}
}}

\section{Methods}{

\describe{
\item{\code{evaluate(v)}}{Evaluate the value of the variable.}

\item{\code{get(k)}}{Retrieve the value of a variable.}

\item{\code{initialize()}}{Create a new registry and initialize vars to all variables from current session.}

\item{\code{load(env_file)}}{Load an environment file and set all exported variables.}

\item{\code{set(k, v)}}{Add a new variable to the registry, and evaluate for references to other variables.}
}}

