% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_python.R
\name{remove_python}
\alias{remove_python}
\title{Disable tracking python dependencies and remove associated files.}
\usage{
remove_python(
  project = NULL,
  python_dir = TRUE,
  requirements_file = TRUE,
  prompt = interactive()
)
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active project will
be used. If no project is currently active, then the current working
directory is used instead.}

\item{python_dir}{Remove the directory containing the Python environment and packages?}

\item{requirements_file}{Remove the Python requirements file?}

\item{prompt}{Prompt user before performing each action?}
}
\value{
The project directory, invisibly. Note that this function is normally
called for its side effects.
}
\description{
\code{remove_python} disables \code{renv} python dependency tracking and removes the
associated directories and files used by \code{renv}.
}
\examples{
\dontrun{
  remove_python()
}


}
