% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv-package.R
\docType{package}
\name{renv-package}
\alias{renv}
\alias{renv-package}
\title{renv}
\description{
Use `renv` to define custom \R environments, which can then be bound to
different projects.
}
\details{
Inspired by the Python `virtualenv` module.
}
