% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{renv_config}
\alias{renv_config}
\title{Define an R Virtual Environment.}
\usage{
renv_config(r_version = getRversion(), r_repos = getOption("repos"),
  r_libs = character(), r_libs_overlay = FALSE)
}
\arguments{
\item{r_version}{\code{character[1]}: The version of \R to be used.}

\item{r_repos}{\code{character[*]}: The \R repositories to use with this project.}

\item{r_libs}{\code{character[*]}: The \R libraries associated with this environment.}

\item{r_libs_overlay}{\code{logical[1]}:   Overlay \code{r_libs} on top of the default \R libraries?}
}
\description{
Provide the definition of a new \R virtual environment. Configurations created
with this function are then passed to \code{\link[=renv_create]{renv_create()}} to create a virtual
environment associated with this configuration.
}
\seealso{
Other renv: \code{\link{renv_activate}},
  \code{\link{renv_create}}, \code{\link{renv_deactivate}},
  \code{\link{renv_edit}}
}
\concept{renv}
