% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_create}
\alias{renv_create}
\title{Create an renv}
\usage{
renv_create(name, config = renv_config())
}
\arguments{
\item{name}{The name of the environment to be created.}

\item{config}{The configuration for this `renv`. See [config]
for more details.}
}
\description{
Create an `renv` environment.
}
