% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_edit}
\alias{renv_edit}
\title{Edit an R Virtual Environment}
\usage{
renv_edit(renv = NULL, local = FALSE)
}
\arguments{
\item{renv}{The name of an \code{renv} environment.}

\item{local}{Boolean; is this a project-local virtual environment?}
}
\description{
Edit the configuration file associated with an \R virtual environment. This
can be useful if you need to manually tweak or edit values in a pre-existing
environment.
}
\seealso{
Other renv: \code{\link{renv_activate}},
  \code{\link{renv_config}}, \code{\link{renv_create}},
  \code{\link{renv_deactivate}}
}
\concept{renv}
