% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_remove}
\alias{renv_remove}
\title{Remove a Virtual Environment}
\usage{
renv_remove(name = NULL)
}
\arguments{
\item{name}{The name of a virtual environment. If \code{NULL}, then the active
virtual environment (if any) will be used.}
}
\description{
Remove an existing virtual environment.
}
