% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{renv_restore}
\alias{renv_restore}
\title{Restore a Virtual Environment}
\usage{
renv_restore(manifest = NULL, confirm = interactive())
}
\arguments{
\item{manifest}{The manifest to be used for restoration of the associated
virtual environment.}

\item{confirm}{Boolean; prompt the user before taking any action?}
}
\description{
Restore a virtual environment from a manifest, as generated by
\code{\link[=renv_snapshot]{renv_snapshot()}}.
}
