% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{renv_snapshot}
\alias{renv_snapshot}
\title{Snapshot a Virtual Environment}
\usage{
renv_snapshot(renv = NULL, file = NULL)
}
\arguments{
\item{file}{The location where the library manifest should be written.
If \code{NULL}, the manifest (as an \R list, with each entry representing
an installed package in the library) is returned instead.}
}
\description{
Snapshot a virtual environment, creating a manifest file that can be used
to later restore the virtual environment.
}
\concept{reproducibility}
