% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-python.R
\name{renv_use_python_virtualenv_impl}
\alias{renv_use_python_virtualenv_impl}
\title{Internal helper for activating a Python virtual environment}
\usage{
renv_use_python_virtualenv_impl(
  project,
  name = NULL,
  version = NULL,
  python = NULL
)
}
\arguments{
\item{project}{The project directory.}

\item{name}{The environment name, if any. If unset, it should be constructed
based on the Python executable used (note: \emph{not} the version parameter)}

\item{version}{The \emph{requested} version of Python (which may not be the actual version!)
This version should be used as a hint for finding an appropriate version
of Python, if the environment needs to be re-created.}

\item{python}{The copy of Python to be used. When unset, an appropriate version of Python
should be discovered based on the \code{version} parameter.}
}
\value{
The path to the Python binary in the associated virtual environment.
}
\description{
Internal helper for activating a Python virtual environment
}
\keyword{internal}
