% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{restore}
\alias{restore}
\title{Restore a Virtual Environment}
\usage{
restore(manifest = NULL, confirm = interactive())
}
\arguments{
\item{manifest}{The manifest to be used for restoration of the associated
virtual environment. When \code{NULL}, the most recently-generated manifest
for this project is used.}

\item{confirm}{Boolean; prompt the user before taking any action?}
}
\description{
Restore a virtual environment from a manifest, as previously generated by
\code{\link[=snapshot]{snapshot()}}.
}
