% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{revert}
\alias{revert}
\title{Revert Lockfile}
\usage{
revert(project = NULL, commit = "HEAD")
}
\arguments{
\item{commit}{The commit associated with a prior version of the lockfile.}
}
\description{
Revert the lockfile to its contents at a prior commit.
}
\details{
The \code{revert()} function is currently only implemented for projects using
\code{git} for version control.
}
