% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox.R
\docType{data}
\name{sandbox}
\alias{sandbox}
\title{Interact with the renv Sandbox}
\format{
An object of class \code{list} of length 4.
}
\usage{
sandbox
}
\description{
Internal tools for interacting with the \code{renv} sandbox.
}
\keyword{datasets}
