% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{save_diagnostics}
\alias{save_diagnostics}
\title{Returns a list with the diagnostics report}
\usage{
save_diagnostics(project = NULL)
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active project will
be used. If no project is currently active, then the current working
directory is used instead.}
}
\value{
A list.
}
\description{
Summarizing the state of a project using renv.
This report can occasionally be useful when diagnosing issues with renv.
}
