% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot}
\alias{snapshot}
\title{Snapshot a Virtual Environment}
\usage{
snapshot(project = NULL, file = file.path(project, "renv.lock"),
  confirm = interactive())
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active
project will be used. If no project has been specifically activated,
the current working directory is used.}

\item{file}{The location to where the generated lockfile should be written.
When \code{NULL}, the lockfile (as an \R object) is returned directly instead.}

\item{confirm}{Boolean; prompt the user before taking any action?}
}
\description{
Snapshot a virtual environment, creating a \strong{lockfile} file that can be used
to later restore the virtual environment. See the \link{lockfile} documentation
for more details on the structure of a lockfile.
}
\seealso{
Other reproducibility: \code{\link{lockfile}},
  \code{\link{restore}}
}
\concept{reproducibility}
