% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use}
\alias{use}
\title{Use a set of Packages}
\usage{
use(..., library = NULL)
}
\arguments{
\item{...}{The \R packages to be used with this script.}

\item{library}{The library path into which the requested packages should be installed.}
}
\value{
This function is normally called for its side effects.
}
\description{
Given a set of \R package requirements, install those packages into the
library path requested via \code{library}, and then activate that library path.
}
\details{
Normally, the call to \code{renv::use()} is generated by \code{\link[=embed]{embed()}}. This gives
a mechanism for users to encode an \code{renv} lockfile directly into an \R
script, such that running that script will then use the packages requested
via the generated \code{renv::use()} call.
}
