% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{use_local}
\alias{use_local}
\title{Toggle between Global and Project-local Virtual Environments}
\usage{
use_local(local)
}
\arguments{
\item{local}{Boolean; should project-local virtual environments be used?}
}
\description{
Toggle between global and project-local virtual environments.
}
\details{
This function is for expert use only -- the typical way to initialize a
project-local environment is with the \code{\link[=init]{init()}} function.

Setting this toggles how the various other \code{renv} APIs work -- for example,
in the following:\preformatted{renv::local(TRUE)
renv::create("test")
renv::activate("test")
}

a project-local virtual environment called \code{"test"} is created, and then
activated and used for that project.
}
