% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-python.R
\name{use_python}
\alias{use_python}
\title{Use Python}
\usage{
use_python(python = NULL, type = "virtualenv", name = NULL, ...,
  project = NULL)
}
\arguments{
\item{python}{The path to a Python binary. This can be the path to a Python
binary on the system, or the path to a Python binary within an
already-existing Python environment. If \code{NULL}, the \code{RETICULATE_PYTHON}
environment variable is checked; if that is not set, then the default
version of \code{python} on the \code{PATH} is used instead.}

\item{type}{The type of Python environment to use. By default, a project-local
virtual environment will be created with the requested version of Python,
as supplied by \code{python}. Ignored if the requested version of \code{python} lives
within a pre-existing Python environment.}

\item{name}{When \code{type} is \code{"virtualenv"} or \code{"conda"}, this argument can be
used to supply the name or path that should be used for the associated
Python environment. Ignored if the requested version of \code{python} lives
within a pre-existing Python environment.}

\item{...}{Optional arguments; currently unused.}

\item{project}{The project directory. If \code{NULL}, then the active
project will be used. If no project has been specifically activated,
the current working directory is used.}
}
\description{
Associate a version of Python with your project. When active, \code{renv} will
take care of capturing your Python dependencies when \code{renv::snapshot()} is called,
and re-installing your Python dependencies when \code{renv::restore()} is called.
}
