% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R, R/deactivate.R
\name{activate}
\alias{activate}
\alias{deactivate}
\title{Activate or Deactivate a Project}
\usage{
activate(project = NULL, profile = NULL)

deactivate(project = NULL)
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active project will
be used. If no project is currently active, then the current working
directory is used instead.}

\item{profile}{The profile to be activated. When \code{NULL}, the default
profile is activated instead. See \code{vignette("profiles", package = "renv")}
for more information.}
}
\value{
The project directory, invisibly. Note that this function is normally
called for its side effects.
}
\description{
\code{activate()} enables \code{renv} for a project in both the current session and
in all future sessions. You should not generally need to call \code{activate()}
yourself as it's called automatically by \code{\link[=init]{init()}}, which is the best
way to start using \code{renv} in a new project.

\code{activate()} does two things. It:
\enumerate{
\item Loads the requested project via \code{\link[=load]{load()}}.
\item Adds \code{source("renv/activate.R")}, the "auto-loader", to the project
\code{.Rprofile}. This script ensures that the correct version of renv is
installed then calls \code{\link[=load]{load()}}, ensuring that future \R sessions
launched from the project automatically load that project, without
explicit user action.
}

\code{deactivate()} removes the infrastructure automatically activate \code{renv} in
new session. This removes the auto-loader from the \code{.Rprofile}; it does
not delete the lockfile or the project library.
}
\examples{
\dontrun{

# activate the current project
renv::activate()

# activate a separate project
renv::activate("~/projects/analysis")

# deactivate the currently-activated project
renv::deactivate()

}
}
\concept{renv}
