% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{activate}
\alias{activate}
\title{Activate a Project}
\usage{
activate(project = NULL, version = NULL)
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active
project will be used. If no project has been specifically activated,
the current working directory is used.}

\item{version}{The version of \code{renv} to associate with this project. When
\code{NULL}, the version recorded in the project's lockfile will be used
(if any); otherwise, the version of \code{renv} currently loaded in the \R
session will be used instead.}
}
\description{
Use \code{activate()} to write the infrastructure needed to ensure that
newly-launched \R projects will load the project's private library on launch,
alongside any other project-specific state recorded for the project.
}
\seealso{
Other renv: \code{\link{deactivate}}
}
\concept{renv}
