% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{activate}
\alias{activate}
\title{Activate an R Virtual Environment}
\usage{
activate(name = NULL, project = NULL, local = NULL)
}
\arguments{
\item{name}{The name of a virtual environment. If \code{NULL}, then the active
virtual environment (if any) will be used.}

\item{project}{The project directory. If \code{NULL}, then the active
project will be used. If no project has been specifically activated,
the current working directory is used.}

\item{local}{Boolean; activate a project-local environment?}
}
\description{
Activate an \R virtual environment. This binds the requested project to the
virtual environment called \code{name}. Newly-launched \R sessions in this project
will use this virtual environment.
}
\seealso{
Other renv: \code{\link{create}}, \code{\link{deactivate}},
  \code{\link{modify}}
}
\concept{renv}
