% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R, R/deactivate.R
\name{activate}
\alias{activate}
\alias{deactivate}
\title{Activate or Deactivate a Project}
\usage{
activate(project = NULL, profile = NULL)

deactivate(project = NULL)
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active project will
be used. If no project is currently active, then the current working
directory is used instead.}

\item{profile}{The profile to be activated. When \code{NULL}, the default
profile is activated instead. See \code{vignette("profiles", package = "renv")}
for more information.}
}
\value{
The project directory, invisibly. Note that this function is normally
called for its side effects.
}
\description{
\code{activate()} enables renv for a project in both the current session and
in all future sessions. You should not generally need to call \code{activate()}
yourself as it's called automatically by \code{\link[=init]{init()}}, which is the best
way to start using renv in a new project.

\code{activate()} first calls \code{\link[=scaffold]{scaffold()}} to set up the project
infrastructure. Most importantly, this creates a project library and adds a
\code{.Rprofile} to ensure that the project library is automatically used for all
future instances of the project. It then calls \code{\link[=load]{load()}} to use the
project library for the current session.

\code{deactivate()} removes the infrastructure automatically activate renv in
new session. This removes the auto-loader from the \code{.Rprofile}; it does
not delete the lockfile or the project library.
}
\examples{
\dontrun{

# activate the current project
renv::activate()

# activate a separate project
renv::activate("~/projects/analysis")

# deactivate the currently-activated project
renv::deactivate()

}
}
