% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{User-Level Configuration of renv}
\description{
Configure different behaviors of \code{renv}.
}
\details{
For a given configuration option:
\enumerate{
\item If an \R option of the form \code{renv.config.<name>} is available,
then that option's value will be used;
\item If an environment variable of the form \code{RENV_CONFIG_<NAME>} is available,
then that option's value will be used;
\item Otherwise, the default for that particular configuration value is used.
}

Any periods (\code{.})s in the option name are transformed into underscores (\code{_})
in the environment variable name, and vice versa. For example, the
configuration option \code{auto.snapshot} could be configured as:
\itemize{
\item \code{options(renv.config.auto.snapshot = <...>)}
\item \code{Sys.setenv(RENV_CONFIG_AUTO_SNAPSHOT = <...>)}
}

Note that if both the R option and the environment variable are defined, the
R option will be used instead. Environment variables can be more useful when
you want a particular configuration to be automatically inherited by child
processes; if that behavior is not desired, then the R option may be
preferred.

If you want to set and persist these options across multiple projects, it is
recommended that you set them in your user startup files (e.g. in
\code{~/.Rprofile}).
}
\section{Configuration}{


The following \code{renv} configuration options are available:

\tabular{lll}{
\strong{Name}        \tab \strong{Type}      \tab \strong{Description} \cr
\code{auto.snapshot} \tab \code{logical(1)}  \tab Automatically snapshot changes to the project library after a call to \code{\link[=install]{install()}}. \cr
\code{use.cache}     \tab \code{logical(1)}  \tab Use the global cache when installing packages?
}
}

