% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{User-Level Configuration of renv}
\description{
Configure different behaviors of \code{renv}.
}
\details{
For a given configuration option:
\enumerate{
\item If an \R option of the form \code{renv.config.<name>} is available,
then that option's value will be used;
\item If an environment variable of the form \code{RENV_CONFIG_<NAME>} is available,
then that option's value will be used;
\item Otherwise, the default for that particular configuration value is used.
}

Any periods (\code{.})s in the option name are transformed into underscores (\code{_})
in the environment variable name, and vice versa. For example, the
configuration option \code{auto.snapshot} could be configured as:
\itemize{
\item \code{options(renv.config.auto.snapshot = <...>)}
\item \code{Sys.setenv(RENV_CONFIG_AUTO_SNAPSHOT = <...>)}
}

Note that if both the R option and the environment variable are defined, the
R option will be used instead. Environment variables can be more useful when
you want a particular configuration to be automatically inherited by child
processes; if that behavior is not desired, then the R option may be
preferred.

If you want to set and persist these options across multiple projects, it is
recommended that you set them in your user startup files (e.g. in
\code{~/.Rprofile} or \code{~/.Renviron}).
}
\section{Configuration}{


The following \code{renv} configuration options are available:

\tabular{llll}{
\strong{Name} \tab \strong{Type} \tab \strong{Default} \tab \strong{Description} \cr

\code{auto.snapshot} \tab \code{logical[1]} \tab \code{FALSE} \tab
Automatically snapshot changes to the project library after a new package
is installed with \code{renv::install()}, or removed with \code{renv::remove()}?
\cr

\code{connect.timeout} \tab \code{integer[1]} \tab \code{20L} \tab
The amount of time to spend (in seconds) when attempting to download a
file. Only used when the \code{curl} downloader is used.
\cr

\code{connect.retry} \tab \code{integer[1]} \tab \code{3L} \tab
The number of times to attempt re-downloading a file, when transient
errors occur. Only used when the \code{curl} downloader is used.
\cr

\code{external.libraries} \tab \code{character[*]} \tab \code{character()} \tab
A character vector of external libraries, to be used in tandem with your
projects. Be careful when using external libraries: it's possible that
things can break within a project if the version(s) of packages used in
your project library happen to be incompatible with packages in your
external libraries; for example, if your project required \code{xyz 1.0} but
\code{xyz 1.1} was present and loaded from an external library. Can also be an
\R function that provides the paths to external libraries. Library paths
will be expanded through \link{.expand_R_libs_env_var} as necessary.
\cr

\code{repos.override} \tab \code{character[n]} \tab \code{NULL} \tab
Override the R package repositories used during \link{restore}. Primarily
useful for deployment / continuous integration, where you might want
to enforce the usage of some set of repositories over what is defined
in \code{renv.lock} or otherwise set by the R session.
\cr

\code{sandbox.enabled} \tab \code{logical[1]} \tab \code{FALSE} \tab
Enable sandboxing for \code{renv} projects? When active, \code{renv} will attempt to
sandbox the system library, preventing user-installed packages in the
system library from becoming available in \code{renv} projects.
\cr

\code{shims.enabled} \tab \code{logical[1]} \tab \code{TRUE} \tab
Should \code{renv} shims be installed on package load? When enabled, \code{renv}
will install its own shims over the functions \code{install.packages()},
\code{update.packages()} and \code{remove.packages()}, delegating these functions
to \code{renv::install()}, \code{renv::update()} and \code{renv::remove()} as
appropriate.
\cr

\code{snapshot.validate} \tab \code{logical[1]} \tab \code{TRUE} \tab
Validate \R package dependencies when calling snapshot? When \code{TRUE},
\code{renv} will attempt to diagnose potential issues in the project library
before creating \code{renv.lock} -- for example, if a package installed in the
project library depends on a package which is not currently installed.
\cr

\code{user.library} \tab \code{logical[1]} \tab \code{FALSE} \tab
Include the user library on the library paths for your projects? Note that
this risks breaking project encapsulation and is not recommended for
projects which you intend to share or collaborate on with other users. See
also the caveats for the \code{external.libraries} option.
\cr

\code{user.profile} \tab \code{logical[1]} \tab \code{TRUE} \tab
Load the user R profile (typically located at \code{~/.Rprofile}) when \code{renv}
is loaded? Consider disabling this if you require extra encapsulation in
your projects; e.g. if your \code{.Rprofile} attempts to load packages that
you might not install in your projects.
\cr

}
}

\section{Project-Local Settings}{


For settings that should persist alongside a particular project, the
various settings available in \link{settings} can be used.
}

\examples{

# disable automatic snapshots
options(renv.config.auto.snapshot = FALSE)

# disable with environment variable
Sys.setenv(RENV_CONFIG_AUTO_SNAPSHOT = "FALSE")

}
