% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{create}
\alias{create}
\title{Create an R Virtual Environment.}
\usage{
create(name, r_version = getRversion(), r_repos = getOption("repos"),
  r_libs = character(), r_libs_overlay = FALSE, local = FALSE,
  overwrite = FALSE)
}
\arguments{
\item{name}{\code{character[1]}: The name of the virtual environment.}

\item{r_version}{\code{character[1]}: The version of \R to be used.}

\item{r_repos}{\code{character[*]}: The \R repositories to use with this project.}

\item{r_libs}{\code{character[*]}: The \R libraries associated with this environment.}

\item{r_libs_overlay}{\code{logical[1]}:   Overlay \code{r_libs} on top of the default \R libraries?}

\item{local}{Boolean; create a project-local virtual environment?}

\item{overwrite}{Boolean; overwrite a pre-existing virtual environment with
the new definition?}
}
\description{
Create a new virtual environment.
}
\seealso{
Other renv: \code{\link{activate}},
  \code{\link{deactivate}}, \code{\link{modify}}
}
\concept{renv}
