% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{deactivate}
\alias{deactivate}
\title{Deactivate an R Virtual Environment}
\usage{
deactivate(project = NULL)
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active
project will be used. If no project has been specifically activated,
the current working directory is used.}
}
\description{
Deactivate the active virtual environment.
}
\seealso{
Other renv: \code{\link{activate}}
}
\concept{renv}
