% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{deactivate}
\alias{deactivate}
\title{Deactivate an R Virtual Environment}
\usage{
deactivate(project = NULL)
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the project associated
with the active virtual environment (if any) will be used.}
}
\description{
Deactivate the active virtual environment.
}
\seealso{
Other renv: \code{\link{activate}}, \code{\link{create}},
  \code{\link{modify}}
}
\concept{renv}
