% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependencies}
\alias{dependencies}
\title{Find R Package Dependencies in a Project}
\usage{
dependencies(path = getwd())
}
\arguments{
\item{path}{The path to a (possibly multi-mode) \R file, or a directory
containing such files.}
}
\description{
Find \R packages used within a project.
}
