% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed.R
\name{embed}
\alias{embed}
\title{Embed a Lockfile}
\usage{
embed(path = NULL, ..., project = NULL)
}
\arguments{
\item{path}{The path to an \R or R Markdown script.}

\item{...}{Unused arguments, reserved for future expansion. If any arguments
are matched to \code{...}, \code{renv} will signal an error.}

\item{project}{The project directory. If \code{NULL}, then the active project will
be used. If no project is currently active, then the current working
directory is used instead.}
}
\description{
Use \code{embed()} to embed a lockfile directly within a file.
}
\details{
This is primarily useful in tandem with \link{run} -- if you call \code{renv::run()}
on a script containing an inline lockfile, \code{renv} will first provision
a library based on that lockfile definition, and then run the script
using that lockfile.
}
