% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equip.R
\name{equip}
\alias{equip}
\title{Install required system libraries}
\usage{
equip()
}
\value{
This function is normally called for its side effects.
}
\description{
Equip your system with libraries commonly-used during compilation of
base and recommended \R packages. This was previously useful with older
versions of R on windows, but is no longer terribly helpful.
}
\examples{
\dontrun{

# download useful build tools
renv::equip()

}
}
\keyword{internal}
