% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrate.R
\name{hydrate}
\alias{hydrate}
\title{Hydrate a Virtual Environment}
\usage{
hydrate(name = NULL, project = NULL)
}
\arguments{
\item{name}{The name of a virtual environment. If \code{NULL}, then the active
virtual environment (if any) will be used.}

\item{project}{The project directory. If \code{NULL}, then the active
project will be used. If no project has been specifically activated,
the current working directory is used.}
}
\description{
Discover package dependencies within a project, and then ensure those
dependencies are all installed in the requested virtual environment.
}
