% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrate.R
\name{hydrate}
\alias{hydrate}
\title{Hydrate a Project}
\usage{
hydrate(packages = NULL, library = NULL, missing = NULL,
  project = NULL)
}
\arguments{
\item{packages}{The set of \R packages to install. When \code{NULL}, the
set of packages as reported by \code{\link[=dependencies]{dependencies()}} is used.}

\item{library}{The \R library to be hydrated. When \code{NULL}, the active
library as reported by \code{.libPaths()} is used.}

\item{missing}{The behavior to choose when one or more of the requested
packages are not available on the system. See \strong{Missing Packages}
for more details.}

\item{project}{The project directory. If \code{NULL}, then the active project will
be used. If no project is currently active, then the current working
directory is used instead.}
}
\description{
Discover the \R packages used within a project, and then install those
packages into the active library. This effectively allows you to clone the
state of your system R libraries for use within a project library.
}
\details{
\code{hydrate()} attempts to re-use packages already installed on your system, to
avoid unnecessary attempts to download and install packages from CRAN.
\code{hydrate()} will attempt to discover \R packages from the following sources
(in order):
\itemize{
\item The user library,
\item The site library,
\item The system library,
\item The \code{renv} cache.
}

If package is discovered in one of these locations, \code{renv} will attempt to
copy or link that package into the requested library as appropriate.
}
\section{Missing Packages}{


The \code{missing} argument controls what happens when your project requests the
hydration of packages which are not currently installed on your system. The
possible values are:

\tabular{ll}{
\strong{Option}   \tab \strong{Action} \cr
\code{"install"}  \tab Install the latest version of missing packages from CRAN. \cr
\code{"ignore"}   \tab Do nothing; missing packages will not be installed. \cr
}

The \R option \code{renv.hydrate.missing} can be used to control the default behavior
of \code{renv::hydrate()}.
}

