% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize a Project}
\usage{
init(project = NULL, force = FALSE)
}
\arguments{
\item{project}{The project directory.}

\item{force}{Boolean; force initialization? By default, \code{renv} will refuse
to initialize the home directory as a project, to defend against accidental
misusages of \code{init()}.}
}
\description{
Discover packages used within the current project, and then initialize a
project-local private \R library with those packages. The currently-installed
versions of any packages in use (as detected within the default R libraries)
are then installed to the project's private library.
}
\details{
The primary steps taken when initializing a new project are:
\enumerate{
\item \R package dependencies are discovered within the \R files used within
the project with \code{\link[=dependencies]{dependencies()}};
\item Discovered packages are copied into the \code{renv} global package cache, so
these packages can be re-used across future projects as necessary;
\item Any missing \R package dependencies discovered are then installed into
the project's private library;
\item A lockfile capturing the state of the project's library is created
with \code{\link[=snapshot]{snapshot()}};
\item The project is activated with \code{\link[=activate]{activate()}}.
}

This mimics the workflow provided by \code{packrat::init()}, but with more
reasonable default behaviors -- in particular, \code{renv} does not attempt
to download and store package sources, and \code{renv} will re-use packages
that have already been installed whenever possible.
}
