% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load}
\alias{load}
\title{Load a Project}
\usage{
load(project = NULL, quiet = FALSE)
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active project will
be used. If no project is currently active, then the current working
directory is used instead.}

\item{quiet}{Boolean; be quiet during load?}
}
\value{
The project directory, invisibly. Note that this function is normally
called for its side effects.
}
\description{
\code{renv::load()} sets the library paths to use a project-local library, and
setting up the system library \link{sandbox}, if needed.

You should not generally need to call \code{renv::load()} yourself, as it's
called automatically by the project auto-loader created by \code{\link[=init]{init()}}/
\code{\link[=activate]{activate()}}. However, if needed, you can use \code{renv::load("<project>")}
to explicitly load an \code{renv} project located at a particular path.
}
\examples{
\dontrun{

# load a project -- note that this is normally done automatically
# by the project's auto-loader, but calling this explicitly to
# load a particular project may be useful in some circumstances
renv::load()

}
}
