% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{load_vars}
\alias{load_vars}
\title{Load environment variables into the current R session}
\usage{
load_vars(env_file)
}
\arguments{
\item{env_file}{A connection object or character string to pass to \code{readLines}.}
}
\value{
Invisibly,  a list object containing all currently set environment variables.
}
\description{
\code{load_vars} reads the contents of an environment file, loads,
evaluates and sets environment variables in the current R session.
}
\examples{
\dontrun{
# read variables from file
load_vars("~/.bashrc")

# acccess via standard call to Sys.getenv
Sys.getenv("foo") # "bar"

# alternatively, access via the return value
vars <- load_vars(".env")
vars$foo # "bar"

}
}
