% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockfile.R
\name{lockfile}
\alias{lockfile}
\title{Lockfiles}
\description{
A \strong{lockfile} records the state of a project at some point in time. In
particular, this implies capturing the \R packages installed (along with
their versions) within the project's private library.
}
\details{
Projects can be restored from a lockfile using the \code{\link[=restore]{restore()}} function.
This implies re-installing packages into the project's private library,
as encoded within the lockfile.

While lockfiles are normally generated and used with \code{\link[=snapshot]{snapshot()}} /
\code{\link[=restore]{restore()}}, they can also hand-edited if so desired. The structure is
similar to that of the Windows-style \code{.ini} file, with some provisioning for
nested sections.

An example lockfile follows:\preformatted{[renv]
Version=0.1.0

[R]
Version=3.5.1
Repositories=
  CRAN=https://cran.rstudio.com

[R/Package/markdown]
Package=markdown
Version=0.9
Source=CRAN
Hash=8515151150d7372bc76e0af15ef5dee0

[R/Package/mime]
Package=mime
Version=0.6
Source=CRAN
Hash=b1e49df8aef896bc8c0b749ef1da5a48
}

The sections used within a lockfile are described next.
}
\section{[renv]}{


Information about the version of \code{renv} used to manage this project.

\tabular{ll}{
\strong{Version}     \tab The version of the \code{renv} package used with this project. \cr
}
}

\section{[R]}{


Properties related to the version of \R associated with this project.

\tabular{ll}{
\strong{Version}      \tab The version of \R used. \cr
\strong{Repositories} \tab The \R repositories used in this project. \cr
}
}

\section{[R/Package/*]}{


Package records, related to the version of an \R package that was installed
at the time the lockfile was generated.

\tabular{ll}{
\strong{Package}      \tab The package name. \cr
\strong{Version}      \tab The package version. \cr
\strong{Library}      \tab The library this package was installed in. \cr
\strong{Source}       \tab The location from which this package was retrieved. \cr
\strong{Hash}         \tab (Optional) A unique hash for this package, used for package caching. \cr
}

Additional remote fields, further describing how the package can be
retrieved from its corresponding source, will also be included as
appropriate (e.g. for packages installed from GitHub).
}

\section{[Python]}{


Metadata related to the version of Python used with this project (if any).

\tabular{ll}{
\strong{Version} \tab The version of Python being used. \cr
\strong{Type}    \tab The type of Python environment being used ("virtualenv", "conda", "system") \cr
\strong{Name}    \tab The (optional) name of the environment being used.
}

Note that the \code{Name} field may be empty. In that case, a project-local Python
environment will be used instead (when not directly using a system copy of Python).
}

\seealso{
Other reproducibility: \code{\link{restore}},
  \code{\link{snapshot}}
}
\concept{reproducibility}
