% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{modify}
\alias{modify}
\title{Open the Lockfile for Editing}
\usage{
modify(project = NULL, load = TRUE)
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active
project will be used. If no project has been specifically activated,
the current working directory is used.}

\item{load}{Boolean; load lockfile-related state after edit?}
}
\description{
Open the Lockfile for Editing
}
