% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{modify}
\alias{modify}
\title{Modify an R Virtual Environment}
\usage{
modify(name = NULL)
}
\arguments{
\item{name}{The name of a virtual environment. If \code{NULL}, then the active
virtual environment (if any) will be used.}
}
\description{
Modify the configuration file associated with an \R virtual environment. This
can be useful if you need to manually tweak or edit values in a pre-existing
environment.
}
\seealso{
Other renv: \code{\link{activate}}, \code{\link{create}},
  \code{\link{deactivate}}
}
\concept{renv}
