% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{paths}
\alias{paths}
\title{Path Customization}
\description{
Customize the paths that \code{renv} uses for global state storage.
}
\details{
By default, \code{renv} collects state into these folders:

\tabular{ll}{
\strong{Platform} \tab \strong{Location} \cr
Linux        \tab \code{~/.local/share/renv} \cr
macOS        \tab \code{~/Library/Application Support/renv} \cr
Windows      \tab \code{\%APPDATA\%/renv} \cr
}

If desired, this path can be adjusted by setting the \code{RENV_PATHS_ROOT}
environment variable. This can be useful if you'd like, for example, multiple
users to be able to share a single global cache.

The various state sub-directories can also be individually adjusted, if so
desired (e.g. you'd prefer to keep the cache of package installations on a
separate volume). The various environment variables that can be set are
enumerated below:

\tabular{ll}{
\strong{Environment Variable} \tab \strong{Description} \cr
\code{RENV_PATHS_ROOT}        \tab The root path used for global state storage. \cr
\code{RENV_PATHS_BOOTSTRAP}   \tab The library path containing bootstrapped \code{renv} installations. \cr
\code{RENV_PATHS_LIBRARY}     \tab The root path containing different \R libraries. \cr
\code{RENV_PATHS_SOURCE}      \tab The path containing downloaded package sources. \cr
\code{RENV_PATHS_BINARY}      \tab The path containing downloaded package binaries. \cr
\code{RENV_PATHS_CACHE}       \tab The path containing cached package installations. \cr
\code{RENV_PATHS_REPOS}       \tab The path containing cached available package information. \cr
\code{RENV_PATHS_EXTSOFT}     \tab (Windows only) The path containing external software needed for compilation of Windows source packages. \cr
}

Note that \code{renv} will append platform-specific and version-specific entries
to the set paths as required. For example, if you have set:

Sys.setenv(RENV_PATHS_CACHE = "/mnt/shared/renv/cache")

then the directory used for the cache will still depend on the \R version
and also cache

/mnt/shared/renv/cache/R-3.5/v2

This ensures that you can set a single \code{RENV_PATHS_CACHE} environment variable
globally without worry that it may cause collisions or errors if multiple
versions of \R needed to interact with the same cache.

If reproducibility of a project is desired on a particular machine, it is
highly recommended that the \code{renv} cache of installed packages + binary
packages is stored, so that packages can be easily restored in the future --
installation of packages from source can often be arduous.

If you want these settings to persist in your project, it is recommended that
you add these to an appropriate \R startup file. For example, these could be
set in:
\itemize{
\item A project-local \code{.Renviron};
\item The user-level \code{.Renviron};
\item A file at \code{$(R RHOME)/etc/Renviron.site}.
}

Please see ?\link{Startup} for more details.
}
