% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{paths}
\alias{paths}
\title{Path Customization}
\description{
Customize the paths that \code{renv} uses for global state storage.
}
\details{
By default, all state is collected into a directory at \code{~/.renv}. If desired,
this path can be adjusted by setting the \code{RENV_PATHS_ROOT} environment
variable. This can be useful if you'd like multiple users to be able to
share a single global cache of installed packages.

The various state sub-directories can also be individually adjusted, if so
desired (e.g. you'd prefer to keep the cache of package installations on a
separate volume). The various environment variables that can be set are
enumerated below:

\tabular{ll}{
\strong{Environment Variable} \tab \strong{Description} \cr
\code{RENV_PATHS_ROOT}        \tab The root path used for global state storage. \cr
\code{RENV_PATHS_BOOTSTRAP}   \tab The library path containing bootstrapped \code{renv} installations. \cr
\code{RENV_PATHS_LIBRARY}     \tab The root path containing different \R libraries. \cr
\code{RENV_PATHS_SOURCE}      \tab The path containing downloaded package sources. \cr
\code{RENV_PATHS_BINARY}      \tab The path containing downloaded package binaries. \cr
\code{RENV_PATHS_CACHE}       \tab The path containing cached package installations. \cr
\code{RENV_PATHS_REPOS}       \tab The path containing cached available package information. \cr
}

If reproducibility of a project is desired on a particular machine, it is
highly recommended that the \code{renv} cache of installed packages + binary
packages is stored, so that packages can be easily restored in the future --
installation of packages from source can often be arduous.

If you want these settings to persist in your project, it is recommended that
you add these to an appropriate \R startup file. For example, these could be
set in:
\itemize{
\item A project-local \code{.Renviron};
\item The user-level \code{.Renviron};
\item A file at \code{etc/Renviron.site}.
}

Please see ?\link{Startup} for more details.
}
