% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{paths}
\alias{paths}
\title{Path Customization}
\description{
Customize the paths that \code{renv} uses for global state storage.
}
\details{
By default, all state is collected into a directory at \code{~/.renv}. This
directory can be adjusted by setting the \code{RENV_PATHS_ROOT} environment
variable. If desired, this path can be adjusted -- this can be useful if
you want to share a set of environments with multiple users.

The various state sub-directories can also be individually adjusted, if so
desired (e.g. you'd prefer to keep the larger 'library' directory on a
separate volume). The various environment variables that can be set are
enumerated below:

\tabular{ll}{
\strong{Description}          \tab \strong{Environment Variable} \cr
\code{RENV_PATHS_ROOT}        \tab The root path used for global state storage. \cr
\code{RENV_PATHS_BOOTSTRAP}   \tab The library path containing bootstrapped \code{renv} installations. \cr
\code{RENV_PATHS_LIBRARY}     \tab The root path containing different \R libraries. \cr
\code{RENV_PATHS_ENVIRONMENT} \tab The path containing \R virtual environment definitions. \cr
\code{RENV_PATHS_CACHE}       \tab The path containing cached package sources, binaries, and installs. \cr
}
}
