% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{refresh}
\alias{refresh}
\title{Refresh the Local Cache of Packages Available on CRAN}
\usage{
refresh()
}
\description{
Query the active CRAN repositories for available packages, and
update the in-memory cache of these packages.
}
\details{
Note that \R also maintains its own on-disk cache of available packages,
which is used by \code{available.packages()}. Calling \code{refresh()} will force
an update of both types of caches. \code{renv} prefers using an in-memory
cache as on occasion the temporary directory can be slow to access (e.g.
when it is a mounted network filesystem).
}
