% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rehash.R
\name{rehash}
\alias{rehash}
\title{Re-Hash Packages in the renv Cache}
\usage{
rehash(cache = paths$cache(), clean = FALSE, confirm = TRUE)
}
\arguments{
\item{cache}{The path to the \code{renv} cache.}

\item{clean}{Boolean; remove files from their previously-cached locations?}

\item{confirm}{Boolean; prompt the user before taking any action?}
}
\description{
Re-hash packages in the \code{renv} cache, ensuring that any previously-cached
packages are copied (or moved) to a new cache location appropriate for this
version of \code{renv}.
}
\details{
This can be useful if the cache scheme has changed in a new version of
\code{renv}, but you'd like to preserve your previously-cached packages.
}
