% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes.R
\name{remote}
\alias{remote}
\title{Resolve a Remote}
\usage{
remote(spec)
}
\arguments{
\item{spec}{A remote specification.}
}
\description{
Given a remote specification, resolve it into an \code{renv} package record that
can be used for download and installation (e.g. with \link{install}).
}
