% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove}
\alias{remove}
\title{Remove Packages}
\usage{
remove(packages)
}
\arguments{
\item{packages}{A character vector of R packages to remove.}
}
\description{
Remove packages currently installed within a project's private library.
}
