% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv-package.R
\docType{package}
\name{renv-package}
\alias{renv}
\alias{renv-package}
\title{renv}
\description{
Virtual environments for \R, inspired by the Python
\href{https://virtualenv.pypa.io/en/latest/}{virtualenv} module.
}
\details{
Use \code{renv} to provide project-specific configuration to your \R projects.
Using \code{renv}, you can associate projects with:
\enumerate{
\item A particular version of \R,
\item A set of library paths,
}

These environments can be shared by multiple projects, and multiple projects
can bind to the same \code{renv}.
}
