% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv-package.R
\docType{package}
\name{renv-package}
\alias{renv}
\alias{renv-package}
\title{Project-local Environments for R}
\description{
Project-local environments for \R, inspired in part by the Python
\href{https://virtualenv.pypa.io/en/latest/}{virtualenv} and
\href{https://docs.python.org/3/library/venv.html}{venv} modules.
}
\details{
You can use \code{renv} to construct isolated, project-local \R libraries.
Each project using \code{renv} will share package installations from a global
cache of packages, helping to avoid wasting disk space on multiple
installations of a package that might otherwise be shared across projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/renv}
  \item Report bugs at \url{https://github.com/rstudio/renv/issues}
}

}
