% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv-package.R
\docType{package}
\name{renv-package}
\alias{renv}
\alias{renv-package}
\title{Project-local Environments for R}
\description{
Project-local environments for \R.
}
\details{
You can use \code{renv} to construct isolated, project-local \R libraries.
Each project using \code{renv} will share package installations from a global
cache of packages, helping to avoid wasting disk space on multiple
installations of a package that might otherwise be shared across projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rstudio.github.io/renv}
  \item Report bugs at \url{https://github.com/rstudio/renv/issues}
}

}
\author{
\strong{Maintainer}: Kevin Ushey \email{kevin@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
