% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_activate}
\alias{renv_activate}
\title{Activate an R Virtual Environment}
\usage{
renv_activate(name, project = NULL)
}
\arguments{
\item{name}{The name of a virtual environment. If \code{NULL}, then the active
virtual environment (if any) will be used.}

\item{project}{The project directory. If \code{NULL}, then the project associated
with the active virtual environment (if any) will be used.}
}
\description{
Activate an \R virtual environment. This binds the requested project to the
virtual environment called \code{name}. Newly-launched \R sessions in this project
will use this virtual environment.
}
\seealso{
Other renv: \code{\link{renv_create}},
  \code{\link{renv_deactivate}}, \code{\link{renv_edit}}
}
\concept{renv}
