% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_activate}
\alias{renv_activate}
\title{Activate an renv}
\usage{
renv_activate(name, project = ".")
}
\arguments{
\item{name}{The name of the environment you'd like to activate.}

\item{project}{The project directory.}
}
\description{
Activate an \code{renv}. This binds a project to an existing \code{renv}
environment definition. Newly-launched \R sessions in this project will
use the \code{renv} as specified by \code{name}.
}
