% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{renv_config}
\alias{renv_config}
\title{Define an renv}
\usage{
renv_config(r_version = getRversion(), r_libs = character(),
  r_libs_overlay = FALSE)
}
\description{
Define an `renv`.
}
\details{
The fields currently supported are:

\tabular{lll}{
\strong{r_version}      \tab \code{character[1]} \tab The \R version to be used for this project.    \cr
\strong{r_libs}         \tab \code{character[n]} \tab The \R libraries to activate for this project. \cr
\strong{r_libs_overlay} \tab \code{logical[1]}   \tab Overlay requested libraries on top of the default R libraries. \cr
}
}
