% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{renv_config}
\alias{renv_config}
\title{Define an renv}
\usage{
renv_config(r_version = getRversion(), r_libs = NULL)
}
\description{
Define an `renv`.
}
\details{
The fields currently supported are:

\tabular{ll}{
\strong{r_version} \tab The \R version to be used for this project.    \cr
\strong{r_libs}    \tab The \R libraries to activate for this project. \cr
}
}
