% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{renv_config}
\alias{renv_config}
\title{Define an renv}
\usage{
renv_config(renv_version = packageVersion("renv"),
  r_version = getRversion(), r_libs = character(),
  r_libs_overlay = FALSE)
}
\description{
Define an \code{renv}. \code{wat}. \code{wat}.
}
\details{
The fields currently supported are:

\tabular{lll}{
\strong{renv_version}   \tab \code{character[1]} \tab The version of the \code{renv} package to be used with this project. \cr
\strong{r_version}      \tab \code{character[1]} \tab The \R version to be used for this project.    \cr
\strong{r_libs}         \tab \code{character[*]} \tab The \R libraries to activate for this project. \cr
\strong{r_libs_overlay} \tab \code{logical[1]}   \tab Overlay requested libraries on top of the default R libraries. \cr
}
}
