% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_create}
\alias{renv_create}
\title{Create an R Virtual Environment.}
\usage{
renv_create(name, r_version = getRversion(),
  r_repos = getOption("repos"), r_libs = character(),
  r_libs_overlay = FALSE)
}
\arguments{
\item{name}{\code{character[1]}: The name of the virtual environment.}

\item{r_version}{\code{character[1]}: The version of \R to be used.}

\item{r_repos}{\code{character[*]}: The \R repositories to use with this project.}

\item{r_libs}{\code{character[*]}: The \R libraries associated with this environment.}

\item{r_libs_overlay}{\code{logical[1]}:   Overlay \code{r_libs} on top of the default \R libraries?}
}
\description{
Create a new virtual environment.
}
\seealso{
Other renv: \code{\link{renv_activate}},
  \code{\link{renv_deactivate}}, \code{\link{renv_edit}}
}
\concept{renv}
