% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_create}
\alias{renv_create}
\title{Create an R Virtual Environment}
\usage{
renv_create(renv, config = renv_config())
}
\arguments{
\item{renv}{The name of an \code{renv} environment.}

\item{config}{The \code{renv} configuration. See \code{\link[=renv_config]{renv_config()}} for more details.}
}
\description{
Create a new \R virtual environment, associated with a configuration as
generated by \code{\link[=renv_config]{renv_config()}}.
}
\seealso{
Other renv: \code{\link{renv_activate}},
  \code{\link{renv_config}}, \code{\link{renv_deactivate}},
  \code{\link{renv_edit}}
}
\concept{renv}
