% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_edit}
\alias{renv_edit}
\title{Edit an R Virtual Environment}
\usage{
renv_edit(name = NULL)
}
\arguments{
\item{name}{The name of a virtual environment. If \code{NULL}, then the active
virtual environment (if any) will be used.}
}
\description{
Edit the configuration file associated with an \R virtual environment. This
can be useful if you need to manually tweak or edit values in a pre-existing
environment.
}
\seealso{
Other renv: \code{\link{renv_activate}},
  \code{\link{renv_create}}, \code{\link{renv_deactivate}}
}
\concept{renv}
