% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_load}
\alias{renv_load}
\title{Load an R Virtual Environment.}
\usage{
renv_load(project = NULL)
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the project associated
with the active virtual environment (if any) will be used.}
}
\description{
Load the \R virtual environment currently associated with a project.
}
\details{
Normally, this is done automatically on session startup by the infrastructure
initialized by \code{\link[=renv_activate]{renv_activate()}} -- users should not need to call this
function directly.
}
