% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{renv_load}
\alias{renv_load}
\title{Load an renv}
\usage{
renv_load(name, project = ".")
}
\arguments{
\item{name}{The name of the environment you'd like to activate.}
}
\description{
Load the `renv` environment called `name`. Normally this is done
automatically when \R sessions are launched, after calling [activate]
to bind a project to a particular `renv` environment.
}
