% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{renv_snapshot}
\alias{renv_snapshot}
\title{Snapshot a Virtual Environment}
\usage{
renv_snapshot(name = NULL, file = "", confirm = interactive())
}
\arguments{
\item{name}{The name of a virtual environment. If \code{NULL}, then the active
virtual environment (if any) will be used.}

\item{file}{The location where the library manifest should be written. If
\code{NULL}, the manifest is returned as an \R object. If \code{""} (the empty
string), a project-local manifest is written to file.}

\item{confirm}{Boolean; prompt the user before taking any action?}
}
\description{
Snapshot a virtual environment, creating a manifest file that can be used
to later restore the virtual environment.
}
\concept{reproducibility}
