% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair.R
\name{repair}
\alias{repair}
\title{Repair a project library}
\usage{
repair(library = NULL, lockfile = NULL, project = NULL)
}
\arguments{
\item{library}{The \R library to be used. When \code{NULL}, the active project
library will be used instead.}

\item{lockfile}{The path to a lockfile (if any). When available, renv
will use the lockfile when attempting to infer the remote associated
with the inaccessible version of each missing package.}

\item{project}{The project directory. If \code{NULL}, then the active project will
be used. If no project is currently active, then the current working
directory is used instead.}
}
\description{
Repair a project library whose cache symlinks have become broken.
renv will attempt to re-install the requisite packages.
}
