% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{restore}
\alias{restore}
\title{Restore a Virtual Environment}
\usage{
restore(project = NULL, lockfile = NULL, confirm = interactive())
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active
project will be used. If no project has been specifically activated,
the current working directory is used.}

\item{lockfile}{The lockfile to be used for restoration of the associated
virtual environment. When \code{NULL}, the most recently-generated lockfile
for this project is used.}

\item{confirm}{Boolean; prompt the user before taking any action?}
}
\description{
Restore a virtual environment from a lockfile, as previously generated by
\code{\link[=snapshot]{snapshot()}}.
}
\seealso{
Other reproducibility: \code{\link{lockfile}},
  \code{\link{snapshot}}
}
\concept{reproducibility}
