% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\docType{data}
\name{settings}
\alias{settings}
\title{Settings}
\format{An object of class \code{list} of length 2.}
\usage{
settings
}
\description{
Define project-local settings that can be used to adjust the behavior of
\code{renv} with your particular project.
}
\section{Settings}{


\describe{

\item{\code{ignored.packages}}{

A vector of packages, which should be ignored when attempting to snapshot
the project's private library. Note that if a package has already been
added to the lockfile, that entry in the lockfile will not be ignored. \cr

}

}
}

\examples{
\dontrun{

# check the 'ignored.packages' option
renv::settings$ignored.packages()

# ignore the 'tidyverse' package in this project
renv::settings$ignored.packages("tidyverse")

}
}
\keyword{datasets}
