% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot}
\alias{snapshot}
\title{Snapshot a Project}
\usage{
snapshot(project = NULL, library = NULL,
  lockfile = file.path(project, "renv.lock"), confirm = interactive())
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active
project will be used. If no project has been specifically activated,
the current working directory is used.}

\item{library}{The \R library to be snapshotted. When \code{NULL}, the project
library associated with the requested project is used.}

\item{lockfile}{The location where the generated lockfile should be written.
When \code{NULL}, the lockfile (as an \R object) is returned directly instead.}

\item{confirm}{Boolean; prompt the user before taking any action?}
}
\description{
Call \code{snapshot()} to create a \strong{lockfile} capturing the state of a project's
\R package dependencies. The lockfile can be used to later restore these
project's dependencies as required. See the \link{lockfile} documentation for more
details on the structure of a lockfile.
}
\details{
In particular, \code{snapshot()} captures all \R packages currently installed
within the project's private library. Dependencies within any other libraries
(currently active on the library paths or not) are not captured.
}
\seealso{
Other reproducibility: \code{\link{lockfile}},
  \code{\link{restore}}
}
\concept{reproducibility}
