% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\title{Status}
\usage{
status(project = NULL, library = NULL, lockfile = NULL)
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active project will
be used. If no project is currently active, then the current working
directory is used instead.}

\item{library}{The path to a library. By default, the project library
associated with the requested project \code{project} is used.}

\item{lockfile}{The path to a lockfile. By default, the project lockfile
(called \code{renv.lock}) is used.}
}
\description{
Report differences between the project's lockfile and the current state of
the private library (if any).
}
