% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysreqs.R
\name{sysreqs}
\alias{sysreqs}
\title{R System Requirements}
\usage{
sysreqs(packages = NULL, project = NULL)
}
\arguments{
\item{packages}{A vector of \R package names. When \code{NULL}
(the default), the project's package dependencies as reported via
\code{\link[=dependencies]{dependencies()}} are used.}

\item{project}{The project directory. If \code{NULL}, then the active project will
be used. If no project is currently active, then the current working
directory is used instead.}
}
\description{
Compute the system requirements (system libraries; operating system packages)
required by a set of \R packages. Currently available for Linux ohnly.
}
\details{
This function relies on the database of package system requirements
maintained by Posit at \url{https://github.com/rstudio/r-system-requirements},
as well as the "meta-CRAN" service at \url{https://crandb.r-pkg.org}.
}
