% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{upgrade}
\alias{upgrade}
\title{Upgrade renv}
\usage{
upgrade(project = NULL, version = NULL, confirm = interactive())
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active
project will be used. If no project has been specifically activated,
the current working directory is used.}

\item{version}{The version of \code{renv} to be installed. When \code{NULL}, the latest
version of \code{renv} as available on GitHub will be installed.}

\item{confirm}{Boolean; confirm upgrade before proceeding?}
}
\description{
Upgrade the version of \code{renv} associated with a project.
}
