% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{upgrade}
\alias{upgrade}
\title{Upgrade renv}
\usage{
upgrade(project = NULL, remote = "rstudio/renv",
  confirm = interactive())
}
\arguments{
\item{project}{The project directory. If \code{NULL}, then the active
project will be used. If no project has been specifically activated,
the current working directory is used.}

\item{remote}{A 'remote' specification, as understood by the
\href{https://cran.r-project.org/package=remotes}{remotes} package. (Note that
not all remote specifications are yet supported by \code{renv}.) See
\url{https://cran.r-project.org/web/packages/remotes/vignettes/dependencies.html}
for more details. When \code{NULL} (the default), the latest version of \code{renv}
available on GitHub is retrieved instead.}

\item{confirm}{Boolean; confirm upgrade before proceeding?}
}
\description{
Upgrade the version of \code{renv} associated with a project.
}
