% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-python.R
\name{use_python}
\alias{use_python}
\title{Use Python}
\usage{
use_python(python = NULL, type = NULL, name = NULL, ...,
  project = NULL)
}
\arguments{
\item{python}{The path to a Python binary. This can be the path to a Python
binary on the system, or the path to a Python binary within an
already-existing Python environment. If \code{NULL}, the \code{RETICULATE_PYTHON}
environment variable is checked; if that is not set, then the default
version of \code{python} on the \code{PATH} is used instead.}

\item{type}{The type of Python environment to use. When \code{"auto"} (the
default), a project-local environment (virtual environments on Linux /
macOS; conda environments on Windows) will be created. Ignored if the
requested version of \code{python} lives within a pre-existing Python
environment.}

\item{name}{The name or path that should be used for the associated Python
environment. If \code{NULL} and \code{python} points to a Python executable living
within a pre-existing virtual environment, that environment will be used.
Otherwise, a project-local environment will be created instead.}

\item{...}{Optional arguments; currently unused.}

\item{project}{The project directory. If \code{NULL}, then the active
project will be used. If no project has been specifically activated,
the current working directory is used.}
}
\description{
Associate a version of Python with your project. When active, \code{renv} will
take care of capturing your Python dependencies when \code{renv::snapshot()} is called,
and re-installing your Python dependencies when \code{renv::restore()} is called.
}
