% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_lockfile.R
\name{validate_lockfile}
\alias{validate_lockfile}
\title{Validate the renv lockfile against a schema}
\usage{
validate_lockfile(
  project = NULL,
  lockfile = NULL,
  lockfile_schema = NULL,
  engine = "ajv",
  greedy = FALSE,
  error = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{lockfile}{Contents of the lockfile, or a filename containing one.}

\item{lockfile_schema}{Contents of the renv schema, or a filename containing a schema.
If not provided, renv's default schema is used.}

\item{engine}{Specify the validation engine to use.
Options are "imjv" (the default; which uses "is-my-json-valid")
and "ajv" (Another JSON Schema Validator).
The latter supports more recent json schema features.}

\item{greedy}{Continue after first error?}

\item{error}{Throw an error on parse failure?
If \code{TRUE}, then the function returns \code{NULL} on success
(i.e., call only for the side-effect of an error on failure, like \code{stopifnot}).}
}
\description{
\code{renv::validate_lockfile()} can be used to validate your renv.lock
against a default or custom schema.
}
\details{
Uses ROpenSci's \href{https://docs.ropensci.org/jsonvalidate/}{\code{jsonvalidate}} package.
}
\examples{
\dontrun{

# validate the project's lockfile
renv::validate_lockfile()
}
}
